LayoutTests/imported/w3c/web-platform-tests/media-source/mediasource-trackdefault-expected.txt


FAIL Test valid audio kinds: type 'audio', language 'en-US', label 'audio label', multiple kinds, byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid audio kinds: type 'audio', language 'en-US', label 'audio label', kind 'alternative', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid audio kinds: type 'audio', language 'en-US', label 'audio label', kind 'descriptions', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid audio kinds: type 'audio', language 'en-US', label 'audio label', kind 'main', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid audio kinds: type 'audio', language 'en-US', label 'audio label', kind 'main-desc', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid audio kinds: type 'audio', language 'en-US', label 'audio label', kind 'translation', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid audio kinds: type 'audio', language 'en-US', label 'audio label', kind 'commentary', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid audio kinds: type 'audio', language 'en-US', label 'audio label', kind '', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid video kinds: type 'video', language 'en-US', label 'video label', multiple kinds, byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid video kinds: type 'video', language 'en-US', label 'video label', kind 'alternative', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid video kinds: type 'video', language 'en-US', label 'video label', kind 'captions', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid video kinds: type 'video', language 'en-US', label 'video label', kind 'main', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid video kinds: type 'video', language 'en-US', label 'video label', kind 'sign', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid video kinds: type 'video', language 'en-US', label 'video label', kind 'subtitles', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid video kinds: type 'video', language 'en-US', label 'video label', kind 'commentary', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid video kinds: type 'video', language 'en-US', label 'video label', kind '', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid text kinds: type 'text', language 'en-US', label 'text label', multiple kinds, byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid text kinds: type 'text', language 'en-US', label 'text label', kind 'subtitles', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid text kinds: type 'text', language 'en-US', label 'text label', kind 'captions', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid text kinds: type 'text', language 'en-US', label 'text label', kind 'descriptions', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid text kinds: type 'text', language 'en-US', label 'text label', kind 'chapters', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test valid text kinds: type 'text', language 'en-US', label 'text label', kind 'metadata', byteStreamTrackID '1' Can't find variable: TrackDefault
FAIL Test mixed valid and invalid audio kinds: type 'audio', language 'en-US', label 'audio label', multiple kinds, byteStreamTrackID '1' assert_throws_js: TrackDefault construction threw an exception function "function () { new TrackDefault(type, language, label, kinds, byteStreamTrackID); }" threw object "ReferenceError: Can't find variable: TrackDefault" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Test mixed valid and invalid video kinds: type 'video', language 'en-US', label 'video label', multiple kinds, byteStreamTrackID '1' assert_throws_js: TrackDefault construction threw an exception function "function () { new TrackDefault(type, language, label, kinds, byteStreamTrackID); }" threw object "ReferenceError: Can't find variable: TrackDefault" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Test mixed valid and invalid text kinds: type 'text', language 'en-US', label 'text label', multiple kinds, byteStreamTrackID '1' assert_throws_js: TrackDefault construction threw an exception function "function () { new TrackDefault(type, language, label, kinds, byteStreamTrackID); }" threw object "ReferenceError: Can't find variable: TrackDefault" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Test invalid 'type' parameter type passed to TrackDefault constructor: type 'invalid type', language 'en-US', label 'label', multiple kinds, byteStreamTrackID '1' assert_throws_js: TrackDefault construction threw an exception function "function () { new TrackDefault(type, language, label, kinds, byteStreamTrackID); }" threw object "ReferenceError: Can't find variable: TrackDefault" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Test invalid 'kinds' parameter type passed to TrackDefault constructor assert_throws_js: TrackDefault construction threw an exception function "function () { new TrackDefault(type, language, label, kinds, byteStreamTrackID); }" threw object "ReferenceError: Can't find variable: TrackDefault" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Test updating the retval of TrackDefault.kinds does not modify TrackDefault.kinds Can't find variable: TrackDefault
